chore: refactor settings component and tests
